requirements.txt
blinker==1.7.0
click==8.1.7
flask==3.0.2
Flask-Cors==4.0.0
importlib-metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
werkzeug==3.0.1
zipp==3.18.1
